[Gold II] Title: 꿈틀꿈틀 호석 애벌레 - 효율성, Time: 408 ms, Memory: 45612 KB, Score: 66 point -BaekjoonHub

+++ 백준/Gold/20181. 꿈틀꿈틀 호석 애벌레 － 효율성/꿈틀꿈틀 호석 애벌레 － 효율성.java
@@ -3,10 +3,10 @@ import java.io.*;
 
 public class Main {
   public static int N;
-  public static int K;
-  public static int[] arr;
-  public static int[] dp;
-  public static ArrayList<ArrayList<int[]>> deps = new ArrayList<>();
+  public static long K;
+  public static long[] arr;
+  public static long[] dp;
+  public static ArrayList<ArrayList<long[]>> deps = new ArrayList<>();
 
   public static void main(String[] args) throws Exception {
     BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -18,8 +18,8 @@ public class Main {
     StringTokenizer st = new StringTokenizer(br.readLine());
     N = Integer.parseInt(st.nextToken());
     K = Integer.parseInt(st.nextToken());
-    arr = new int[N];
-    dp = new int[N+1];
+    arr = new long[N];
+    dp = new long[N+1];
     for(int i = 0 ; i<=N ; i++){
       deps.add(new ArrayList<>());
     }
@@ -31,25 +31,25 @@ public class Main {
 
   public static void pro(){
     int j =0;
-    int sum = 0;
+    long sum = 0;
     for(int i=0 ; i< N; i++){
       while(j < N && sum < K){
         sum += arr[j];
         j++;
       }
       if(sum >= K){
-        deps.get(j).add(new int[]{i, sum-K});
+        deps.get(j).add(new long[]{i, sum-K});
       }
       sum -= arr[i];
     }
 
     for(int i=1 ; i<=N ; i++){
-      int curMax = 0;
-      for(int[] dep : deps.get(i)){
+      long curMax = 0;
+      for(long[] dep : deps.get(i)){
         if(dep[0]==0){
           curMax = Math.max(curMax, dep[1]);
         }else{
-          curMax = Math.max(curMax, dp[dep[0]]+dep[1]);
+          curMax = Math.max(curMax, dp[(int)dep[0]]+dep[1]);
         }
       }
       dp[i] =  Math.max(dp[i-1], curMax);
